caduceus: create primaryHandler, a router that handles all the primary route path (1) authorizating a request and then (2) notification "notify"
